Split pack out into its own module and used cargo_clap to get more cargo-like CLI args
