+++ src/bin/cargo-wapm.rs
@@ -24,5 +24,6 @@ fn main() -> Result<(), Error> {
 #[derive(Debug, Parser)]
 #[clap(name = "cargo", bin_name = "cargo", version, author)]
 enum Cargo {
+    #[clap(alias = "wasmer")]
     Wapm(Publish),
 }
